loanNotificationApp/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
bcrypt
websockets
pyjwt